Initial browser window size fixes timeout
When using the default browser window size for firefox, the scroll list
used for the initial pull refuses to
scroll on while being traversed by the script, causing the pull to
timeout at the last item visible.
Setting the window to a slightly
larger initial size makes that problem disappear for us. *Browser
window size to fail might vary depending on personal font settings for
the browser*